Change common to namespacesless
