+++ AppInsightsCommon/Constants.ts
@@ -1,7 +1,1 @@
-module Microsoft.ApplicationInsights.Common {
-
-    "use strict";
-
-    ///<summary>Function that returns property name which will identify that monitoring for given instance of XmlHttpRequest is disabled</summary>
-    export var DisabledPropertyName: string = "Microsoft_ApplicationInsights_BypassAjaxInstrumentation";
-}
+export const DisabledPropertyName: string = "Microsoft_ApplicationInsights_BypassAjaxInstrumentation";
