+++ AppInsightsCommon/applicationinsights-common.ts
@@ -1,0 +1,2 @@
+export import CommonUtil = require("./Util");
+export {_InternalMessageId as _InternalMessageIdFoo} from './Enums';
